requirements.txt
# Core HTTP and async libraries
httpx==0.27.0
aiohttp==3.9.1
requests==2.31.0

# Database
aiosqlite==0.19.0

# Gemini Flash (replaces xAI + OpenRouter)
google-generativeai>=0.8.0

# Data handling and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.12.0

# Environment variables
python-dotenv==1.0.0

# Pydantic and data validation
pydantic==2.8.2
pydantic-settings

# Logging and monitoring
structlog==23.2.0

# Type hints
typing-extensions>=4.11

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# Rate limiting
ratelimit==2.2.1

# Configuration management
pyyaml==6.0.1

# WebSocket streaming (future: predict.fun real-time)
websockets>=12.0

# News & sentiment pipeline
feedparser>=6.0.0

# JSON repair (for LLM output parsing)
json-repair

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
